% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_apply.R
\name{sliding_apply}
\alias{sliding_apply}
\title{Apply a window function across a time series}
\usage{
sliding_apply(
  data,
  column = "accel",
  window_size = 1000,
  fun,
  output_name = "result",
  ...
)
}
\arguments{
\item{data}{A data.frame containing the time series.}

\item{column}{Name of the column to apply the window function to.}

\item{window_size}{Integer window size (=size of the sliding window->number of points).}

\item{fun}{The window function to apply (e.g., window_mean).}

\item{output_name}{Name of the new column to store results (output column to create).}

\item{...}{Additional arguments to pass to the window function.}
}
\value{
The same data frame with a new column containing results.
}
\description{
Apply a window function across a time series
}
\examples{
\dontrun{
df <- load_accel_csv("Example_window.csv")
df <- sliding_apply(df, "accel", 1000, window_mean, "mean")
}

}
