% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{window_binary_trend}
\alias{window_binary_trend}
\title{Binary trend in a window}
\usage{
window_binary_trend(x)
}
\arguments{
\item{x}{Numeric vector (representing a window)}
}
\value{
+1, -1, or 0
}
\description{
Returns:
\itemize{
\item +1 if the value increases over the window
\item -1 if it decreases
\item 0 if no change
}
}
\examples{
window_binary_trend(c(1,2,3))
window_binary_trend(c(3,2,1))

}
