% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{window_mean}
\alias{window_mean}
\title{Mean value of a window}
\usage{
window_mean(x)
}
\arguments{
\item{x}{Numeric vector (a window from the time series)}
}
\value{
Mean of x
}
\description{
Mean value of a window
}
\examples{
window_mean(1:5)
}
