% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{window_threshold_change}
\alias{window_threshold_change}
\title{Threshold change within a window}
\usage{
window_threshold_change(x, threshold)
}
\arguments{
\item{x}{Numeric vector (representing a window)}

\item{threshold}{Numeric threshold value}
}
\value{
A numeric value (positive, negative (Change value), or 0)

@examples
window_threshold_change(c(1, 1.8, 1.4), threshold = 0.5)
}
\description{
Checks how much the values increase or decrease in the window.
Returns:
\itemize{
\item largest positive change (if above threshold),
\item or largest negative change (if above threshold),
\item or 0 if no change exceeds threshold.
}
}
